uweflix/requirements.txt
asgiref==3.6.0
autopep8==2.0.1
Django==4.1.4
django-environ==0.9.0
psycopg2==2.9.5
pycodestyle==2.10.0
sqlparse==0.4.3
tomli==2.0.1
tzdata==2022.7
django-crispy-forms>=1.14.0
pillow==9.4.0
djangorestframework==3.14.0
djoser==2.1.0